Move the controller channel to a dedicated module
